Updated homepage, added search functionality, and food database
Linked search page to the icon skylar has created. Created look.dart that houses the search page. Food.dart contains the meal information that look.dart pulls from. It will also be the foundation for the filter page to be implemented at a later date. I DO NOT think we should merge this branch to main yet, as i could not integrate our current map page onto this and it was causing me many conflicts.